activate result delete sql building tests and fix data type bug

+++ src/test/resources/db/value/value_delete_user.sql
@@ -1,3 +1,4 @@
-PREPARE value_delete_user (text) AS
+PREPARE value_delete_user (text, bigint) AS
     DELETE FROM user_values
-     WHERE group_id = $1;
+     WHERE group_id = $1
+       AND user_id = $2;
