+++ src/test/resources/db/value/value_delete_user_mysql.sql
@@ -1,3 +1,4 @@
 PREPARE value_delete_user FROM
    'DELETE FROM user_values
-     WHERE group_id = ?';
+     WHERE group_id = ?
+       AND user_id = ?';
